requirements.txt
flask
requests
python-dotenv
numpy
supabase
langchain-core